A layout for Dvorak gamers
Some games don't let you remap all the keys (especially backspace). This
layout uses a modified version of the Kenesis Dvorak key layout as the
default layout and then has a layout with a right handed keypad and a
Qwerty layout where the left hand is shifted one character right so you
can leave your left hand on the home keys and still do w for forward, a
for left, s for back and f for right. It also puts space for the
backspace key so you can use it for jump.